Move the DirectoryMirror class to a Filesystem namespace
